rename Template -> sbkTemplate
