+++ build.gradle
@@ -354,6 +354,9 @@ dependencies {
     api project(':driver-cephs3')
     api project(':driver-openio')
     api project(':driver-leveldb')
+
+    /* below line is a signature */
+    /* api project(':driver-sbktemplate') */
 }
 
 //create a single Jar with all dependencies
