+++ settings.gradle
@@ -59,4 +59,5 @@ include 'driver-cephs3'
 include 'driver-openio'
 include 'driver-leveldb'
 
-
+/* below line is a signature */
+/* include 'driver-sbktemplate' */
